change /recent to use new route
