+++ src/lib/discord/image-utils.ts
@@ -4,6 +4,12 @@ import {
   editDiscordMessageWithImage,
   getRatingComment
 } from './responses';
+import { prepareCreditData } from '@/server/services/credit-data';
+import { ImageCache, renderLastCreditImage } from '@/lib/render-image';
+import { fetchImageForServer } from '@/lib/render-image-server';
+import { Image, loadImage } from 'skia-canvas';
+import { getRatingImageUrl } from '@/lib/rating-calculator';
+import { DIFFICULTY_ENUM } from '@/lib/db/types';
 
 export interface SnapshotData {
   publicId: string;
@@ -122,5 +128,218 @@ export async function generateAndSendProfileImage({
     // Fallback to regular message if image generation fails
     console.error('Error generating image:', imageError);
     await editDiscordMessage(applicationId, interactionToken, { embeds: [embedData], components });
+  }
+}
+
+export interface CreditImageOptions {
+  userId: string;
+  discordUserId: string;
+  region: 'intl' | 'jp';
+  applicationId: string;
+  interactionToken: string;
+  skip?: number;
+}
+
+export async function generateAndSendCreditImage({
+  userId,
+  discordUserId,
+  region,
+  applicationId,
+  interactionToken,
+  skip = 0,
+}: CreditImageOptions): Promise<void> {
+  const regionName = region === 'jp' ? 'Japan' : 'International';
+
+  try {
+    // Show loading message
+    await editDiscordMessage(applicationId, interactionToken, {
+      embeds: [{
+        title: `🔄 Loading ${regionName} Recent Plays`,
+        description: `<@${discordUserId}> Generating your recent play image...`,
+        color: DISCORD_COLORS.BLURPLE,
+        fields: [{
+          name: '📊 Status',
+          value: '⏳ Generating Image',
+          inline: false,
+        }],
+        footer: {
+          text: 'tomomai ともマイ • maimai DX score tracker',
+        },
+        timestamp: new Date().toISOString(),
+      }],
+    });
+
+    // Calculate beforeDate for pagination
+    // We need to fetch data to get the beforeDate for the skip+1 credit
+    let beforeDate: Date | undefined = undefined;
+
+    // If skip > 0, we need to find the playedAt of the credit we want to skip to
+    if (skip > 0) {
+      // Fetch credits to find the one we want
+      let currentSkip = 0;
+      let tempBeforeDate: Date | undefined = undefined;
+
+      while (currentSkip < skip) {
+        const tempResult = await prepareCreditData(userId, region, tempBeforeDate);
+        if (tempResult.type === "error" || !tempResult.hasPreviousCredit) {
+          // No more credits available
+          throw new Error('No more plays found');
+        }
+        // Set beforeDate to just before this credit to get the next older one
+        tempBeforeDate = new Date(tempResult.credit.playedAt.getTime() - 1);
+        currentSkip++;
+      }
+
+      beforeDate = tempBeforeDate;
+    }
+
+    // Prepare credit data
+    const prepareDataResult = await prepareCreditData(userId, region, beforeDate);
+
+    if (prepareDataResult.type === "error") {
+      await editDiscordMessage(applicationId, interactionToken, {
+        embeds: [{
+          title: '📊 No Recent Plays Found',
+          description: skip === 0
+            ? `You don't have any recent ${regionName} region plays yet!`
+            : `No more plays found.`,
+          color: DISCORD_COLORS.YELLOW,
+          footer: {
+            text: 'tomomai ともマイ • maimai DX score tracker',
+          },
+        }],
+      });
+      return;
+    }
+
+    const { credit, snapshot, hasNextCredit, hasPreviousCredit } = prepareDataResult;
+
+    // Cache all images needed for rendering
+    const imagesToCache = [
+      "https://maimaidx.jp/maimai-mobile/img/music_dx.png",
+      "https://maimaidx.jp/maimai-mobile/img/music_standard.png",
+      getRatingImageUrl(snapshot.rating),
+      snapshot.iconUrl,
+      snapshot.classRankUrl,
+      snapshot.courseRankUrl,
+      `/res/trophy/normal.png`,
+      `/res/trophy/bronze.png`,
+      `/res/trophy/silver.png`,
+      `/res/trophy/gold.png`,
+      `/res/trophy/rainbow.png`,
+      `/res/character/${snapshot.gameVersion}.png`,
+      `/res/logo/${snapshot.gameVersion}.png`,
+      `/res/bg/${snapshot.gameVersion}.png`,
+      `/res/bg/${snapshot.gameVersion}_long.png`,
+      `/res/badge/${snapshot.gameVersion}/none.png`,
+      `/res/badge/${snapshot.gameVersion}/sync.png`,
+      `/res/badge/${snapshot.gameVersion}/fc.png`,
+      `/res/badge/${snapshot.gameVersion}/fc+.png`,
+      `/res/badge/${snapshot.gameVersion}/fs.png`,
+      `/res/badge/${snapshot.gameVersion}/fs+.png`,
+      `/res/badge/${snapshot.gameVersion}/fdx.png`,
+      `/res/badge/${snapshot.gameVersion}/fdx+.png`,
+      ...['percentage_blue', 'percentage_red', 'percentage_gold',
+        'score_blue', 'score_red', 'score_gold', 'score_big_blue', 'score_big_red', 'score_big_gold',
+        'score_num_count', 'score_num_count_big',
+        'level_basic', 'level_advanced', 'level_expert', 'level_master', 'level_remaster']
+        .map(path => `/res/numbers/${path}.png`),
+      ...['score_table', 'fast_late', 'track_1', 'track_2', 'track_3',
+        'dxscore', 'star_1', 'star_2', 'star_3']
+        .map(path => `/res/songs/${path}.png`),
+      ...['base', 'sync_base', 'sync', 'fc_base', 'fc', 'fc+', 'ap_base', 'ap', 'ap+',
+        'fs_base', 'fs', 'fs+', 'fdx_base', 'fdx', 'fdx+']
+        .map(path => `/res/icons/${path}.png`),
+      ...Object.values(DIFFICULTY_ENUM).map(difficulty => `/res/songs/song_${difficulty}.png`),
+      ...Object.values(DIFFICULTY_ENUM).map(difficulty => `/res/songs/music_jacket_${difficulty}.png`),
+      ...credit.tracks.map(s => s.cover),
+    ];
+
+    const cache: ImageCache = {};
+    await Promise.all(
+      imagesToCache.map(async (url) => {
+        try {
+          if (url.startsWith('data:')) return;
+          return fetchImageForServer(url).then(async img => {
+            let memo: Image | null = null;
+            cache[url] = async () => memo || (memo = await loadImage(img));
+          });
+        } catch (error) {
+          console.warn(`⚠️ Failed to cache image: ${url}`);
+        }
+      })
+    );
+
+    // Render the image
+    const canvas = await renderLastCreditImage(credit, snapshot, cache);
+    const imageBuffer = Buffer.from(await canvas.toBuffer('jpg', { density: 2, quality: 0.9 }));
+
+    // Create embed data
+    const embedData = {
+      title: `🎵 ${regionName} Recent Plays`,
+      description: `<@${discordUserId}> Here are your recent plays!`,
+      color: DISCORD_COLORS.BLURPLE,
+      fields: [
+        {
+          name: '📅 Played At',
+          value: `<t:${Math.floor(credit.playedAt.getTime() / 1000)}:R>`,
+          inline: true,
+        },
+        {
+          name: '🎮 Tracks',
+          value: credit.tracks.length.toString(),
+          inline: true,
+        },
+      ],
+      footer: {
+        text: 'tomomai ともマイ • maimai DX score tracker',
+      },
+      timestamp: new Date().toISOString(),
+    };
+
+    // Create navigation buttons
+    const components = [
+      {
+        type: 1, // ACTION_ROW
+        components: [
+          {
+            type: 2, // BUTTON
+            custom_id: `recents_${discordUserId}_${region}_${skip - 1}`,
+            label: 'Newer Play',
+            style: 2, // SECONDARY
+            emoji: {
+              name: '⬅️'
+            },
+            disabled: !hasNextCredit,
+          },
+          {
+            type: 2, // BUTTON
+            custom_id: `recents_${discordUserId}_${region}_${skip + 1}`,
+            label: 'Older Play',
+            style: 2, // SECONDARY
+            emoji: {
+              name: '➡️'
+            },
+            disabled: !hasPreviousCredit,
+          },
+        ],
+      },
+    ];
+
+    // Send the image
+    await editDiscordMessageWithImage(applicationId, interactionToken, embedData, imageBuffer, components);
+
+  } catch (error) {
+    console.error('Error generating credit image:', error);
+    await editDiscordMessage(applicationId, interactionToken, {
+      embeds: [{
+        title: '❌ Error',
+        description: `Failed to generate image: ${error instanceof Error ? error.message : 'Unknown error'}`,
+        color: DISCORD_COLORS.RED,
+        footer: {
+          text: 'tomomai ともマイ • maimai DX score tracker',
+        },
+      }],
+    });
   }
 }
