+++ src/lib/discord/index.ts
@@ -24,5 +24,5 @@ export { handleAlbumPreferenceSelection } from './commands/album-preference';
 export type { AlbumPreferenceOptions } from './commands/album-preference';
 
 // Image generation utilities
-export { generateAndSendProfileImage } from './image-utils';
-export type { ImageGenerationOptions, SnapshotData } from './image-utils';
+export { generateAndSendProfileImage, generateAndSendCreditImage } from './image-utils';
+export type { ImageGenerationOptions, SnapshotData, CreditImageOptions } from './image-utils';
